Corrections to challenge 5, changes to the data type of the point class

+++ Reto-5-proyect.punto1/shape.py
@@ -2,20 +2,20 @@ import math
 
 
 class Point:
-    def __init__(self, x: int, y: int):
+    def __init__(self, x: float, y: float):
         self._x = x
         self._y = y
 
-    def get_x(self) -> int:
+    def get_x(self) -> float:
         return self._x
 
-    def set_x(self, new_x: int):
+    def set_x(self, new_x: float):
         self._x = new_x
 
-    def get_y(self) -> int:
+    def get_y(self) -> float:
         return self._y
 
-    def set_y(self, new_y: int):
+    def set_y(self, new_y: float):
         self._y = new_y
 
     def compute_distance(self, other_point: "Point") -> float:
